move type aliases to core/types.py
